refactor(Avatar): public api 추가 및 경로 변경
