+++ src/shared/design-system/avatar/index.ts
@@ -1,0 +1,3 @@
+export { default as Avatar } from "./avatar"
+export { default as AvatarImage } from "./avatar-image"
+export { default as AvatarFallback } from "./avatar-fallback"
